Add star multiplier param in worlddata Add itemget star reward plist and add star when itemget DeadUI achievement exclamation sign implementation Replace store button by achievement button in MainMenu Start the game immediately after clicking the run button Show unlocked UI in the achievement view if it is unlocked

+++ DumplingJump/DU/Classes/GameUI.h
@@ -33,7 +33,7 @@
 -(void) fadeOut;
 -(void) resetUI;
 -(void) updateScore:(int)score;
--(void) updateStar:(int)starNum;
+-(void) updateStar:(float)starNum;
 -(void) addStageClearMessageWithDistance:(int) distance;
 -(void) addAchievementUnlockMessageWithName:(NSString *)name;
 -(void) showRebornButton;
